+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Show clearer error if "rules" is missing from config.
+- Allow ignore parameters for the `forbid-field` rule.
 
 ## 0.0.16 (2023-05-31)